# Improve performance of fqcnt_jl1_klib.jl