@lh3 2020-05-21
Yes, it is a significant improvement on uncompressed fastq. Do you know why julia is not that fast on compressed fastq? Also, which libz.so does the script is calling? Is it the system zlib (which is 1.2.7 on CentOS7)? I have updated the timing in README.md. I will update fqcnt/README.md and blog post tomorrow.

The more interesting benchmark is bedcov. It is simpler than the first. I don't understand why Julia can't get good performance. Thanks.