@bicycle1885 2020-05-21
Recent Julia distributions are shipped with zlib.so and it is used if you write `ccall((func, "libz"), ...)` by default. Unfortunately, that zlib.so is very slow because the build script was not configured correctly (in fact, this is my bad). That problem was fixed recently (yes, [very recently!](https://github.com/JuliaPackaging/Yggdrasil/pull/1051)) but the fix is not yet included even in the latest release (i.e., Julia 1.4.1). So, in order to fix the problem right now, you may need to use your own libz.so installed in your system, for example.

This is an example patch on my machine:
```patch
diff --git a/lib/Klib.jl b/lib/Klib.jl
index 1b7fc39..a8f8d21 100644
--- a/lib/Klib.jl
+++ b/lib/Klib.jl
@@ -120,6 +120,9 @@ end
 
 tostring(b::ByteBuffer, n::Int) = unsafe_string(b.a, n)
 
+const libz = "/usr/lib/x86_64-linux-gnu/libz.so"
+#const libz = "libz"
+
 #
 # GzFile
 #
@@ -127,17 +130,17 @@ mutable struct GzFile <: IO
        fp::Ptr{Cvoid}
 
        function GzFile(fn::String, mode = "r")
-               x = ccall((:gzopen, "libz"), Ptr{Cvoid}, (Cstring, Cstring), fn, mode)
+               x = ccall((:gzopen, libz), Ptr{Cvoid}, (Cstring, Cstring), fn, mode)
                y = x == C_NULL ? nothing : new(x)
                if y != nothing finalizer(Base.close, y) end
                return y
        end
 end
 
-Base.readbytes!(fp::GzFile, buf::Vector{UInt8}) = ccall((:gzread, "libz"), Cint, (Ptr{Cvoid}, Ptr{Cvoid}, Cuint), fp.fp, buf, length(buf))
+Base.readbytes!(fp::GzFile, buf::Vector{UInt8}) = ccall((:gzread, libz), Cint, (Ptr{Cvoid}, Ptr{Cvoid}, Cuint), fp.fp, buf, length(buf))
 
 function Base.close(fp::GzFile)
-       ret = fp.fp != C_NULL ? ccall((:gzclose, "libz"), Cint, (Ptr{Cvoid},), fp.fp) : -1
+       ret = fp.fp != C_NULL ? ccall((:gzclose, libz), Cint, (Ptr{Cvoid},), fp.fp) : -1
        fp.fp = C_NULL
        return ret
 end
```

And this is a comparison before the patch:
```
~/w/biofast (master|✚1…) $ hyperfine 'fqcnt/fqcnt_jl1_klib.jl biofast-data-v1/M_abscessus_HiSeq.fq.gz'
Benchmark #1: fqcnt/fqcnt_jl1_klib.jl biofast-data-v1/M_abscessus_HiSeq.fq.gz
  Time (mean ± σ):     13.930 s ±  0.058 s    [User: 13.710 s, System: 0.683 s]
  Range (min … max):   13.859 s … 14.023 s    10 runs
```

and after the patch:
```
~/w/biofast (master|✚1…) $ hyperfine 'fqcnt/fqcnt_jl1_klib.jl biofast-data-v1/M_abscessus_HiSeq.fq.gz'
Benchmark #1: fqcnt/fqcnt_jl1_klib.jl biofast-data-v1/M_abscessus_HiSeq.fq.gz
  Time (mean ± σ):      7.555 s ±  0.037 s    [User: 7.325 s, System: 0.707 s]
  Range (min … max):    7.504 s …  7.616 s    10 runs
 ```

I will check bedcov, too. But maybe I have not enough time to improve it immediately.