@KristofferC 2020-05-21
>  That problem was fixed recently (yes, very recently!) but the fix is not yet included even in the latest release

Do you know if the upcoming 1.5 has the improved version? I would guess no, since the fix to Yggdrasil was merged so recently. We should definitely try get that in, in my opinion.