@bicycle1885 2020-05-21
I didn't know, but I expected ;) Considering @lh3 will update his blog post very soon and using the nightly build is not recommended in general, I think the quickest fix we can take for benchmarking is to hardcode the zlib path as I've shown above. Of course, [Zlib_jll.jl](https://github.com/JuliaBinaryWrappers/Zlib_jll.jl) may be used instead, but it adds a new dependency in the benchmark.