@lh3 2020-05-21
I have updated README and blog post using the system zlib. Julia is on par with faster languages now. Note that Klib.jl in this repo still uses "libz". When timing, I was using the system zlib on CentOS7.